Removed the 'get' methods, which aren't used, and aren't necessary.
The methods in Plot and Trace overrode the base Dictionary class
method, changing the default from None to an empty string, thinking
that would be more useful for Jinja templates. However, the methods
are not used in the way envisioned and since the methods cause
confusion they have been removed.